Factor out container from resource
