+++ web-bundle/src/main/java/com/graphhopper/http/GraphHopperBundle.java
@@ -32,7 +32,7 @@ import com.graphhopper.GraphHopperAPI;
 import com.graphhopper.http.health.GraphHopperHealthCheck;
 import com.graphhopper.http.health.GraphHopperStorageHealthCheck;
 import com.graphhopper.jackson.Jackson;
-import com.graphhopper.reader.gtfs.GraphHopperGtfs;
+import com.graphhopper.reader.gtfs.PtRouteResource;
 import com.graphhopper.reader.gtfs.GtfsStorage;
 import com.graphhopper.reader.gtfs.PtEncodedValues;
 import com.graphhopper.resources.*;
@@ -194,7 +194,7 @@ public class GraphHopperBundle implements ConfiguredBundle<GraphHopperBundleConf
         final GHDirectory ghDirectory = new GHDirectory(configuration.get("graph.location", "target/tmp"), DAType.RAM_STORE);
         final GtfsStorage gtfsStorage = GtfsStorage.createOrLoad(ghDirectory);
         EncodingManager encodingManager = PtEncodedValues.createAndAddEncodedValues(EncodingManager.start()).add(new CarFlagEncoder()).add(new FootFlagEncoder()).build();
-        final GraphHopper graphHopperStorage = GraphHopperGtfs.createOrLoad(encodingManager, gtfsStorage, configuration);
+        final GraphHopper graphHopperStorage = PtRouteResource.createOrLoad(encodingManager, configuration);
         final TranslationMap translationMap = new TranslationMap().doImport();
         final LocationIndex locationIndex = graphHopperStorage.getLocationIndex();
         environment.jersey().register(new AbstractBinder() {
@@ -210,7 +210,7 @@ public class GraphHopperBundle implements ConfiguredBundle<GraphHopperBundleConf
             }
         });
         environment.jersey().register(NearestResource.class);
-        environment.jersey().register(GraphHopperGtfs.class);
+        environment.jersey().register(PtRouteResource.class);
         environment.jersey().register(new PtIsochroneResource(gtfsStorage, encodingManager, graphHopperStorage.getGraphHopperStorage(), locationIndex));
         environment.jersey().register(I18NResource.class);
         environment.jersey().register(InfoResource.class);
